Subindo relacionamento e comunicação onibus - parada

+++ src/main/java/br/com/muriel/busIOT/rest/dto/BusStopDTO.java
@@ -4,13 +4,15 @@ import br.com.muriel.busIOT.rest.enums.Direction;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class BusStopDTO {
 
     @JsonIgnore
     private Long id;
     private String name;
-    private Double latitude;
-    private Double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     private Direction direction;
 }
